Work on jupyter notebooks. Included download path for data.
